requirements.txt
torch>=1.8.0
torchvision==0.19.0
torchaudio==2.4.0
transformers==4.42.0
numpy>=1.19.0
pandas>=1.1.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn==0.13.0
tqdm>=4.45.0
pyyaml>=5.1
swanlab>=0.3.0
python-dotenv==1.0.0
tensorboard>=2.5.0
pytorch-lightning>=1.5.0
biopython==1.83
colorama==0.4.6
pytest==7.4.0
omegaconf==2.3.0 